// src/src/api/json-rpc/types/enums/JSONRPCErrorCode.ts
export enum JSONRPCErrorCode {
    PARSE_ERROR = -32700,
    INVALID_REQUEST = -32600,
    METHOD_NOT_FOUND = -32601,
    INVALID_PARAMS = -32602,
    INTERNAL_ERROR = -32603,
    SERVER_ERROR = -32000,
    APPLICATION_ERROR = -32099,
    SYSTEM_ERROR = -32098,
    TRANSPORT_ERROR = -32097,
    GENERIC_ERROR = -32096,
}

export enum JSONRPCErrorHttpCodes {
    PARSE_ERROR = 500,
    INVALID_REQUEST = 400,
    METHOD_NOT_FOUND = 404,
    INVALID_PARAMS = 400,
    INTERNAL_ERROR = 500,
    SERVER_ERROR = 500,
    APPLICATION_ERROR = 500,
    SYSTEM_ERROR = 500,
    TRANSPORT_ERROR = 500,
    GENERIC_ERROR = 500,
}